% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{beta_loglik}
\alias{beta_loglik}
\title{Log-likelihood function for beta-coalescent model}
\usage{
beta_loglik(t, alpha)
}
\arguments{
\item{t}{Tree}

\item{alpha}{Alpha parameter}
}
\value{
Log-likelihood
}
\description{
Log-likelihood function for beta-coalescent model
}
