% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_psize}
\alias{beta_psize}
\title{Size distribution for Beta-coalescent}
\usage{
beta_psize(n, alpha)
}
\arguments{
\item{n}{Number of lineages}

\item{alpha}{Alpha parameter}
}
\value{
Distribution of size of the coalescence event
}
\description{
Size distribution for Beta-coalescent
}
