% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta_simtree}
\alias{beta_simtree}
\title{Simulation from Beta-coalescent model}
\usage{
beta_simtree(n, alpha)
}
\arguments{
\item{n}{Number of lineages}

\item{alpha}{Alpha parameter}
}
\value{
Simulated phylogeny
}
\description{
Simulation from Beta-coalescent model
}
