% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{compare_simu}
\alias{compare_simu}
\title{Comparison of simulated and calculated coalescent probabilities}
\usage{
compare_simu(
  rt = 1.5,
  vt = 10,
  nt = 10,
  nt1 = 20,
  nrep = 1000,
  method = "multinomial"
)
}
\arguments{
\item{rt}{Reproduction number}

\item{vt}{Variance of offspring distribution (if equal to R_t then Poisson is used otherwise NegBin)}

\item{nt}{Population size at time t}

\item{nt1}{Population size at time t+1}

\item{nrep}{Number of repeats to simulate}

\item{method}{Which method to use for simulation, options are: rejection or multinomial}
}
\value{
Vector of inclusive coalescent probabilities for number of lines from 1 to nt1
}
\description{
Comparison of simulated and calculated coalescent probabilities
}
