% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_mixed_poisson.R
\name{compare_simu_mix_pois_2}
\alias{compare_simu_mix_pois_2}
\title{Comparison of simulated and calculated coalescent probabilities}
\usage{
compare_simu_mix_pois_2(
  lambda,
  q,
  nt = 10,
  nt1 = 20,
  n = 15,
  type = "inclusive",
  nrep = 1000,
  method = "multinomial"
)
}
\arguments{
\item{nt}{Population size at time t}

\item{nt1}{Population size at time t+1}

\item{n}{Number of observed lineages}

\item{type}{Either 'inclusive' or 'exclusive'}

\item{nrep}{Number of repeats to simulate}

\item{method}{Which method to use for simulation, options are: rejection or multinomial}

\item{rt}{Reproduction number}

\item{vt}{Variance of offspring distribution (if equal to R_t then Poisson is used otherwise NegBin)}
}
\value{
Vector of inclusive coalescent probabilities for number of lines from 1 to nt1
}
\description{
Comparison of simulated and calculated coalescent probabilities
}
