% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EpiLambda}
\alias{EpiLambda}
\title{EpiLambda: Epidemic lambda-coalescent model}
\description{
EpiLambda is a R package that implements the epidemic lambda-coalescent model.

The main functions of the package are:
\itemize{

\item TODO
}
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
