% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda_loglik}
\alias{lambda_loglik}
\title{Log-likelihood function for lambda-coalescent model}
\usage{
lambda_loglik(t, lambda)
}
\arguments{
\item{t}{Tree}

\item{lambda}{Lambda function of (n,k)}
}
\value{
Log-likelihood
}
\description{
Log-likelihood function for lambda-coalescent model
}
