% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda_simtree}
\alias{lambda_simtree}
\title{Simulation from lambda-coalescent model}
\usage{
lambda_simtree(n, lambda)
}
\arguments{
\item{n}{Number of lineages}

\item{lambda}{Lambda function of (n,k)}
}
\value{
Simulated phylogeny
}
\description{
Simulation from lambda-coalescent model
}
