% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{negbin_inclusive}
\alias{negbin_inclusive}
\title{Inclusive probability of coalescence of k lineages when the offspring distribution is Negative-Binomial}
\usage{
negbin_inclusive(k, nt, r, log = F)
}
\arguments{
\item{k}{Number of lineages}

\item{nt}{Population size at time t}

\item{r}{Dispersion parameter of Negative-Binomial}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Inclusive probability of coalescence of k lineages when the offspring distribution is Negative-Binomial
}
