% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{negbin_p2t}
\alias{negbin_p2t}
\title{Probability of coalescence of two lineages when the offspring distribution is Negative-Binomial}
\usage{
negbin_p2t(nt, r)
}
\arguments{
\item{nt}{Population size at time t}

\item{r}{Dispersion parameter of Negative-Binomial}
}
\value{
Probability of coalescence
}
\description{
Probability of coalescence of two lineages when the offspring distribution is Negative-Binomial
}
