% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{negbin_pnt}
\alias{negbin_pnt}
\title{Probability of coalescence of n lineages when the offspring distribution is Negative-Binomial}
\usage{
negbin_pnt(n, nt, r, log = F)
}
\arguments{
\item{n}{Number of lineages}

\item{nt}{Population size at time t}

\item{r}{Dispersion parameter of Negative-Binomial}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Probability of coalescence of n lineages when the offspring distribution is Negative-Binomial
}
