% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{new_loglik}
\alias{new_loglik}
\title{Log-likelihood function for new lambda-coalescent model}
\usage{
new_loglik(t, nt, r)
}
\arguments{
\item{t}{Tree}

\item{nt}{Population size parameter}

\item{r}{Dispersion parameter}
}
\value{
Log-likelihood
}
\description{
Log-likelihood function for new lambda-coalescent model
}
