% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{new_mle}
\alias{new_mle}
\title{MLE for new lambda-coalescent model}
\usage{
new_mle(t, nt)
}
\arguments{
\item{t}{Tree}

\item{nt}{Optional, if provided only the mle of r will be computed}
}
\value{
Log-likelihood
}
\description{
MLE for new lambda-coalescent model
}
