% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{new_simtree}
\alias{new_simtree}
\title{Simulation from new lambda-coalescent model}
\usage{
new_simtree(n, nt, r)
}
\arguments{
\item{n}{Number of lineages}

\item{nt}{Population size parameter}

\item{r}{Dispersion parameter}
}
\value{
Simulated phylogeny
}
\description{
Simulation from new lambda-coalescent model
}
