% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{omega_psize}
\alias{omega_psize}
\title{Size distribution for omega-coalescent model}
\usage{
omega_psize(n, nt, r)
}
\arguments{
\item{n}{Number of lineages}

\item{nt}{Population size parameter}

\item{r}{Dispersion parameter}
}
\value{
Distribution of size of the coalescence event
}
\description{
Size distribution for omega-coalescent model
}
