% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{omega_simtree}
\alias{omega_simtree}
\title{Simulation from omega-coalescent model}
\usage{
omega_simtree(n, nt, r)
}
\arguments{
\item{n}{Number of lineages}

\item{nt}{Population size parameter}

\item{r}{Dispersion parameter}
}
\value{
Simulated phylogeny
}
\description{
Simulation from omega-coalescent model
}
