% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{pois_exclusive}
\alias{pois_exclusive}
\title{Exclusive probability of coalescence of k lineages when the offspring distribution is Poisson}
\usage{
pois_exclusive(n, k, nt, log = F)
}
\arguments{
\item{n}{Number of observed lineages}

\item{k}{Number of lineages to coalesce}

\item{nt}{Population size at time t}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Exclusive probability of coalescence of k lineages when the offspring distribution is Poisson
}
