% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{pois_inclusive}
\alias{pois_inclusive}
\title{Inclusive probability of coalescence of k lineages when the offspring distribution is Poisson}
\usage{
pois_inclusive(k, nt, log = F)
}
\arguments{
\item{k}{Number of lineages to coalesce}

\item{nt}{Population size at time t}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Inclusive probability of coalescence of k lineages when the offspring distribution is Poisson
}
