% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{pois_p2t}
\alias{pois_p2t}
\title{Probability of coalescence of two lineages when the offspring distribution is Poisson}
\usage{
pois_p2t(nt, log = F)
}
\arguments{
\item{nt}{Population size at time t}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Probability of coalescence of two lineages when the offspring distribution is Poisson
}
