% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simul_inclusive}
\alias{simul_inclusive}
\title{Simulation of inclusive coalescent probabilities using either rejection sampling or multinomial sampling}
\usage{
simul_inclusive(
  rt,
  vt = rt,
  nt,
  nt1 = round(rt * nt),
  nrep = 1000,
  method = "rejection"
)
}
\arguments{
\item{rt}{Reproduction number}

\item{vt}{Variance of offspring distribution (if equal to R_t then Poisson is used otherwise NegBin)}

\item{nt}{Population size at time t}

\item{nt1}{Population size at time t+1}

\item{nrep}{Number of repeats to simulate}

\item{method}{Which method to use for simulation, options are: rejection or multinomial}
}
\value{
Vector of inclusive coalescent probabilities for number of lines from 1 to nt1
}
\description{
Simulation of inclusive coalescent probabilities using either rejection sampling or multinomial sampling
}
