% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simul_pcoal}
\alias{simul_pcoal}
\title{Simulation of coalescent probabilities using either rejection sampling or multinomial sampling}
\usage{
simul_pcoal(
  rt,
  vt = rt,
  nt,
  nt1 = round(rt * nt),
  type = "inclusive",
  n,
  nrep = 1000,
  method = "rejection"
)
}
\arguments{
\item{rt}{Reproduction number}

\item{vt}{Variance of offspring distribution (if equal to R_t then Poisson is used otherwise NegBin)}

\item{nt}{Population size at time t}

\item{nt1}{Population size at time t+1}

\item{type}{Either inclusive or exclusive}

\item{n}{Number of lineages observed at time t}

\item{nrep}{Number of repeats to simulate}

\item{method}{Which method to use for simulation, options are: rejection or multinomial}
}
\value{
Vector of inclusive coalescent probabilities for number of lines from 1 to n
}
\description{
Simulation of coalescent probabilities using either rejection sampling or multinomial sampling
}
