% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simul_pnt}
\alias{simul_pnt}
\title{Simulation of p_nt using either rejection sampling or multinomial sampling}
\usage{
simul_pnt(
  rt,
  nt,
  vt = rt,
  nt1 = round(rt * nt),
  nrep = 1000,
  max = 10,
  method = "rejection"
)
}
\arguments{
\item{rt}{Reproduction number}

\item{nt}{Population size at time t}

\item{vt}{Variance of offspring distribution (if equal to R_t then Poisson is used otherwise NegBin)}

\item{nt1}{Population size at time t+1}

\item{nrep}{Number of repeats to simulate}

\item{max}{Maximum value of n to consider in p_nt}

\item{method}{Which method to use for simulation, options are: rejection or multinomial}
}
\value{
Vector of p_nt values for n from 1 to max
}
\description{
Simulation of p_nt using either rejection sampling or multinomial sampling
}
