% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{beta_lambda}
\alias{beta_lambda}
\title{Lambda function for Beta-coalescent}
\usage{
beta_lambda(n, k, alpha)
}
\arguments{
\item{n}{Number of lineages}

\item{k}{Number of lineages to coalesce}

\item{alpha}{Alpha parameter}
}
\value{
Rate of coalescence
}
\description{
Lambda function for Beta-coalescent
}
