% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{mixed_pois_2_exclusive}
\alias{mixed_pois_2_exclusive}
\title{Exclusive probability of coalescence of k lineages when the offspring distribution is Poisson}
\usage{
mixed_pois_2_exclusive(n, k, nt, nt1, lambda, q, log = F)
}
\arguments{
\item{n}{Number of observed lineages}

\item{k}{Number of lineages to coalesce}

\item{nt}{Population size at time t}

\item{nt1}{Population size at time t+1}

\item{lambda}{2-dimensional vector of Poisson rates}

\item{q}{Probability of each emmission rate}

\item{log}{Whether to return the log of the probability}
}
\value{
Probability of coalescence
}
\description{
Exclusive probability of coalescence of k lineages when the offspring distribution is Poisson
}
